# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Vilo AI API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface CheckoutInput {
  userName: string;
  userEmail: string;
  /** @nullable */
  userPhone?: string | null;
}

export interface CheckoutResult {
  url: string;
}

export type ViloUserStatus = typeof ViloUserStatus[keyof typeof ViloUserStatus];


export const ViloUserStatus = {
  pending: 'pending',
  active: 'active',
  failed: 'failed',
} as const;

export interface ViloUser {
  id: number;
  name: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  viloNumber?: string | null;
  status: ViloUserStatus;
  createdAt: string;
}

export interface ErrorResponse {
  error: string;
}