# lib/api-zod/src/generated/types/checkoutInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Vilo AI API
 * OpenAPI spec version: 0.1.0
 */

export interface CheckoutInput {
  userName: string;
  userEmail: string;
  /** @nullable */
  userPhone?: string | null;
}