# lib/api-zod/src/generated/types/viloUser.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Vilo AI API
 * OpenAPI spec version: 0.1.0
 */
import type { ViloUserStatus } from './viloUserStatus';

export interface ViloUser {
  id: number;
  name: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  viloNumber?: string | null;
  status: ViloUserStatus;
  createdAt: string;
}